requirements.txt
BeautifulSoup==3.2.1
Django==1.5.5
Pillow==2.1.0
South==0.8.2
beautifulsoup4==4.3.1
boto==2.18.0
cmsplugin-filer==0.9.5
cmsplugin-zinnia==0.4
dj-database-url==0.2.2
dj-static==0.0.5
django-admin-jqueryui==1.8.22
django-blog-zinnia==0.12.3
django-ckeditor==4.0.2
django-classy-tags==0.4
django-filer==0.9.5
django-inline-ordering==1.0.2
django-mailchimp-v1.3==1.3.0
django-mptt==0.5.2
django-polymorphic==0.5.1
django-reversion==1.7.1
django-sekizai==0.7
django-storages==1.1.8
django-tagging==0.3.1
django-toolbelt==0.0.1
django-xmlrpc==0.1.5
djangocms-admin-style==0.1.4
djangocms-text-ckeditor==2.0.1
easy-thumbnails>=1.4
gunicorn==18.0
html5lib==1.0b3
psycopg2==2.5.1
pyparsing==1.5.7
python-dateutil==2.1
pytz==2013b
six==1.3.0
static==0.4
wsgiref==0.1.2
https://github.com/divio/django-cms/archive/develop.zip
